# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  responses:
...
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Unprocessible:
      description: HTTP Unprocessible Entity 422
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    VerificationResult:
      type: object
      properties:
        verificationTimestamp:
          type: string
          format : date-time
        lifecycleStatus: #!! is not known when verifying external SDs -> Julius: You are supposed to compute the status (active, invalid, etc.) from the available information. That is the result of the validation. Maybe rename to just status
          type: string # does the self-description have a lifecycle status in this catalogue?
          description: Lifecycle-state an SD would get, if it would be added to the catalogue.
        issuer: #!! Is this the participant? -> Julius: Yes, the participant that created the SD
          type: string # contains the id of the issuers self-description if available  #!! NOT FURTHER SPECIFIED
          description: ID of the issues Self-Description, if available.
        issuedDateTime:
          type: string
          format: date-time
        validatorDids:
          type: array
          items:
            type: string
            description: The dids of validators, having signed parts of the Self-Description
      required:
        - verificationTimestamp
        - lifecycleStatus #!! is not known when verifying external SDs. Remove from required -> Julius: You are supposed to compute the status (active, invalid, etc.) from the available information. That is the result of the validation. Maybe rename to just status
        - issuer #!! Comment on issuer: if available, but still required?? Remove from required -> Julius: required
        - issuedDateTime
        - validatorDids
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /verification:
    get:
      tags:
        - Verification
      summary: Show a HTML page to verify (portions of) a signed Self-Description
      operationId: verifyPage
      responses:
        '200':
          description: >-
            HTML document that contains a query field to verify (portions of)
            Self-Descriptions.
          content:
            text/html:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ClientError'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Verification
      summary: >-
        Send a JSON-LD document to verify with the information from the
        Catalogue
      operationId: verify
      parameters:
        - name: verifySemantics
          in: query
          description: 'To verify SD semantics or not'
          required: false
          schema:
            type: boolean
            default: true
        - name: verifySchema
          in: query
          description: 'To verify SD agains schema or not'
          required: false
          schema:
            type: boolean
            default: true  
        - name: verifySignatures
          in: query
          description: 'To verify SD signatures or not'
          required: false
          schema:
            type: boolean
            default: true  
      requestBody:
        description: >-
          JSON-LD document to be verified object to send queries. Use
          "application/json" for openCypher queries. A Catalogue may also
          support the other content   types  depending   on   its   supported  
          query  languages   but   only "application/json" is mandatory.
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
        '400':
          $ref: '#/components/responses/ClientError'
        '422':
          $ref: '#/components/responses/Unprocessible'
        '500':
          $ref: '#/components/responses/ServerError'
...